+++ src/algo3/algocity/model/UnidadOcupable.java
@@ -2,11 +2,11 @@ package algo3.algocity.model;
 
 public class UnidadOcupable extends Unidad implements Reparable {
 
-	final int ESTADOINICIAL = 100;
+	final double ESTADOINICIAL = 100;
 
 	int capacidad; // capacidad habitacional
 	int ocupacion;
-	int porcentajeDanios;
+	double porcentajeDanios;
 
 	public UnidadOcupable() {
 
@@ -28,7 +28,7 @@ public class UnidadOcupable extends Unidad implements Reparable {
 		return this.ocupacion;
 	}
 
-	public int getDanios() {
+	public double getDanios() {
 		return porcentajeDanios;
 	}
 
@@ -48,7 +48,7 @@ public class UnidadOcupable extends Unidad implements Reparable {
 		return (this.consultarDisponibilidad() >= cantidad);
 	}
 
-	public int getSalud() {
+	public double getSalud() {
 		return (this.ESTADOINICIAL - this.porcentajeDanios);
 	}
 
@@ -63,7 +63,7 @@ public class UnidadOcupable extends Unidad implements Reparable {
 	// TODO VER si debemos saber antes si cada unidad se
 	// sabe destruir a si misma o el "danio" lo recibe de manera externa
 	// y como destruye un terremoto en un principio
-	public void aplicarDanio(int cantidad) {
+	public void aplicarDanio(double cantidad) {
 		if (this.porcentajeDanios > 100) {
 			this.porcentajeDanios = 100;
 		} else {
@@ -84,7 +84,7 @@ public class UnidadOcupable extends Unidad implements Reparable {
 		}
 	}
 
-	protected int porcentajeReparacion() {
+	protected double porcentajeReparacion() {
 		return (this.ESTADOINICIAL * 10) / 100;
 	}
